refactor(frontend): move chat tag perspective and language types into feature modules
