+++ frontend/src/modules/chats/domain/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./chat.types";
