+++ frontend/src/modules/chats/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./domain";
