+++ frontend/src/modules/languages/domain/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./language.types";
