+++ frontend/src/modules/languages/domain/language.types.ts
@@ -1,0 +1,13 @@
+interface LanguageMeta {
+   type: "language";
+   workspaceOrigin: string;
+   docId: string;
+   versionCid: string;
+   timestamp: string;
+   creatorType: string;
+   language: string;
+}
+
+export interface LanguageRequest {
+   meta: LanguageMeta;
+}
