+++ frontend/src/modules/languages/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./domain";
