+++ frontend/src/modules/perspectives/domain/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./perspective.types";
