+++ frontend/src/modules/perspectives/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./domain";
