+++ frontend/src/modules/tags/domain/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./tag.types";
