+++ frontend/src/modules/tags/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./domain";
